Fix and document rand() & Co.
The C standard says that rand() must return `int', but the algorithm
used had once been written with 32-bit ints in mind.  So fix all that
was broken, notably reduce RAND_MAX to 0x7FFF since this is all we can
do with 16-bit ints.  For the not-so-standard rand_r(), still a
pointer to an unsigned long (as opposed to an unsigned int) is required
to store the current context, since anything else would make no sense
given the 32-bit algorithm used.

Remove the traces of a rand_r() test program that was of no use to the
AVR platform anyway.

Add (#ifdef'ed out) the older poorer de-facto PNRG `standard', just
for reference.

Add a second file random.c, implementing a complementary set of
functions with `rand' replaced by `random', operating on 32-bit
integers (long int).  The algorithm used is identical, only the
result type is different.


git-svn-id: svn://svn.savannah.nongnu.org/avr-libc/trunk/avr-libc@141 1084037a-409f-4a99-8160-1a4b3577d4c1