Make the mode element in the bootloader optional.
Some recent ATtiny devices (apparently starting with the ATtiny2313)
have self-programming support, yet the model is much simpler than it
used to be on the ATmegas.  The entire flash is NRWW, and there is no
separate bootloader area that can be selected.  So do not insist on
having a "mode" element in the bootloader tag.


git-svn-id: svn://svn.savannah.nongnu.org/avr-libc/trunk/avr-libc@865 1084037a-409f-4a99-8160-1a4b3577d4c1